segregated storage
originally writing to an i2c register would always write
to in-memory storage and only write to EEPROM if the bit
was set. Now they are completely distict. Your write to either
one or the other but not both.
This allows for total control of both areas so that programs
can be stored completely before transitioning from in-memory
to EEPROM.